Enable S rule
