+++ connectors/sources/dropbox.py
@@ -64,7 +64,7 @@ else:
 
 
 class EndpointName(Enum):
-    ACCESS_TOKEN = "access_token"
+    ACCESS_TOKEN = "access_token"  # noqa S105
     PING = "ping"
     CHECK_PATH = "check_path"
     FILES_FOLDERS = "files_folders"
