+++ connectors/utils.py
@@ -210,14 +210,15 @@ def convert_to_b64(source, target=None, overwrite=False):
             version = int(platform.mac_ver()[0].split(".")[0])
             # MacOS 13 has changed base64 util
             if version >= 13:
-                cmd = f"{_BASE64} -i {source} -o {temp_target}"
+                cmd = [_BASE64, "-i", source, "-o", temp_target]
             else:
-                cmd = f"{_BASE64} {source} > {temp_target}"
+                cmd = [_BASE64, source, ">", temp_target]
         else:
             # In Linuces, avoid line wrapping
-            cmd = f"{_BASE64} -w 0 {source} > {temp_target}"
+            cmd = [_BASE64, "-w", "0", source, ">", temp_target]
         logger.debug(f"Calling {cmd}")
-        subprocess.check_call(cmd, shell=True)
+        # TODO: make it more robust
+        subprocess.check_call(cmd)  # noqa S603
     else:
         # Pure Python version
         with open(source, "rb") as sf, open(temp_target, "wb") as tf:
@@ -406,25 +407,6 @@ class ConcurrentTasks:
         """Cancels all tasks"""
         for task in self.tasks:
             task.cancel()
-
-
-def get_event_loop(uvloop=False):
-    if uvloop:
-        # activate uvloop if lib is present
-        try:
-            import uvloop
-
-            asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
-        except Exception:
-            pass
-    try:
-        loop = asyncio.get_running_loop()
-    except RuntimeError:
-        loop = asyncio.get_event_loop_policy().get_event_loop()
-        if loop is None:
-            loop = asyncio.new_event_loop()
-            asyncio.set_event_loop(loop)
-    return loop
 
 
 class RetryStrategy(Enum):
@@ -638,7 +620,9 @@ def get_pem_format(key, postfix="-----END CERTIFICATE-----"):
 
 def hash_id(_id):
     # Collision probability: 1.47*10^-29
-    return hashlib.md5(_id.encode("utf8")).hexdigest()
+    # S105 rule considers this code unsafe, but we're not using it for security-related
+    # things, only to generate pseudo-ids for documents
+    return hashlib.md5(_id.encode("utf8")).hexdigest()  # noqa S105
 
 
 def truncate_id(_id):
